requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
fpdf2>=2.7.9
Flask>=2.3.0
pypdf>=4.0.0
sentence-transformers>=2.2.0
pandas>=2.1.0
openpyxl>=3.1.5
reportlab>=3.6.0
